# :thread: (llm): optimize launch