<!--
Thank you for your contribution! 👍
Please make sure to read CONTRIBUTING.md if you have not already. Pull Requests that do not comply with the rules will be arbitrarily closed.
-->

### ✅ Checklist

<!-- Pull Requests must pass the CI and be code reviewed. Set as Draft if the PR is not ready. -->

- [ ] `npx changeset` was attached.
- [ ] **Covered by automatic tests.** <!-- if not, please explain. (Feature must be tested / Bug fix must bring non-regression) -->
- [ ] **Impact of the changes:** <!-- Please take some time to list the impact & what specific areas Quality Assurance (QA) should focus on -->
  - ...

### 📝 Description


## ⚡ Launch Performance Comparison in dev

| Thread                     | Before        | Before 2      | After 1       | After 2       |
| ----------------------- | ------------- | ------------- | ------------- | ------------- |
| getBle                  | 53ms          | 54ms          | 56ms          | 60ms          |
| getSettings             | 70ms          | 66ms          | 57ms          | 61ms          |
| listCachedCurrencyIds   | 103ms         | 85ms          | 57ms          | 62ms          |
| hydrateCurrency         | 412ms, 441ms  | 452ms, 416ms  | 415ms, 424ms  | 451ms, 460ms  |
| listSupportedFiats      | 451ms, 484ms  | 531ms, 500ms  | 415ms, 424ms  | 452ms, 460ms  |
| getsupportedCountervalues | 454ms, 485ms | 532ms, 501ms  | 415ms, 425ms  | 452ms, 460ms  |
| importSettings          | 454ms, 485ms  | 532ms, 501ms  | 415ms, 425ms  | 452ms, 461ms  |
| getAccounts             | 557ms, 516ms  | 598ms, 567ms  | 450ms, 446ms  | 453ms, 488ms  |
| getPostOnboardingState  | 698ms, 616ms  | 666ms, 634ms  | 488ms, 485ms  | 491ms, 524ms  |
| getMarketState          | 715ms, 749ms  | 681ms, 654ms  | 488ms, 486ms  | 492ms, 524ms  |
| getTrustchainState      | 732ms, 765ms  | 698ms, 666ms  | 489ms, 486ms  | 492ms, 524ms  |
| getWalletExportState    | 748ms, 782ms  | 719ms, 683ms  | 489ms, 486ms  | 492ms, 524ms  |
| getProtect              | 765ms, 799ms  | 731ms, 700ms  | 489ms, 486ms  | 492ms, 524ms  |
| getCountervalues        | 800ms, 834ms  | 767ms         | 489ms, 486ms  | 492ms, 524ms  |
| onInitFinished          | 858ms         | 791ms         | 510ms         | 551ms         |


### ❓ Context

- **JIRA or GitHub link**: <!-- Attach the relevant ticket number if applicable. (e.g., [JIRA-123] for Jira or #123 for a Github issue) -->


---

### 🧐 Checklist for the PR Reviewers

<!-- Please do not edit this if you are the PR author -->

- **The code aligns with the requirements** described in the linked JIRA or GitHub issue.
- **The PR description clearly documents the changes** made and explains any technical trade-offs or design decisions.
- **There are no undocumented trade-offs**, technical debt, or maintainability issues.
- **The PR has been tested** thoroughly, and any potential edge cases have been considered and handled.
- **Any new dependencies** have been justified and documented.
- **Performance** considerations have been taken into account. (changes have been profiled or benchmarked if necessary)